checkerboard_wf: JWildfire's GPU snippet drops the int cast of random(_max_checks + 1) — the checker 'sides' were smeared across the board instead of sitting on checker boundaries (overrides.ts patch, registry regenerated: two lines). Found through the _sh8 solid fixture, where post_curl3D amplified the resulting z offset: corr 0.87 → 0.99 (ratio 0.96, blkMAE 2.3); its DOF-less variant 0.89 → 0.99. Port README.

+++ src/core/variations.jwf.ts
@@ -15794,11 +15794,11 @@ if (((min(max(${p[8]}, 0.0), 1.0) == 0) || (abs(${p[4]}) < 0.000001))) {
   if (((_max_checks > 0) && (rnd(rs) < _side_prob))) {
     (*cp) = min(max(${p[7]}, 0.0), 1.0);
     if ((rnd(rs) < 0.5)) {
-      x = ((rnd(rs) * f32((_max_checks + 1))) * ${p[3]});
+      x = (f32(i32((rnd(rs) * f32((_max_checks + 1))))) * ${p[3]});
       y = rnd(rs);
     } else {
       x = rnd(rs);
-      y = ((rnd(rs) * f32((_max_checks + 1))) * ${p[3]});
+      y = (f32(i32((rnd(rs) * f32((_max_checks + 1))))) * ${p[3]});
     }
     z = (${p[4]} * rnd(rs));
   } else {
